requirements.txt
openai-whisper==20230918
torch==2.1.0